rename typescript/index.ts
